Refactor provivion script

+++ provision.sh
@@ -3,39 +3,70 @@
 set -eu
 set -o pipefail
 
-DISK_DEVICE='/dev/disk/by-path/pci-0000:00:14.0-scsi-0:0:2:0'
-WORKSPACE="/home/ubuntu/workspace"
-DOCKER_DIR="${WORKSPACE}/docker_state"
-
-sudo -u ubuntu mkdir -p "$WORKSPACE"
-
-if ! mount "$DISK_DEVICE" "$WORKSPACE" >/dev/null 2>&1 ; then
-  mkfs.ext4 "$DISK_DEVICE"
-  mount "$DISK_DEVICE" "$WORKSPACE"
-  chown ubuntu.ubuntu "$WORKSPACE"
-fi
-
-mkdir -p "${DOCKER_DIR}"
-ln -s "${DOCKER_DIR}" /var/lib/docker || true
-
-apt-get -y update
-apt-get -y clean
-
-apt-get install -y git vim-nox jq cgroup-lite build-essential ntp htop docker.io
-
-wget -qO- https://redirector.gvt1.com/edgedl/go/go1.9.2.linux-amd64.tar.gz | tar -C /usr/local -xzf -
-
-#Set up $GOPATH and add go executables to $PATH
-cat > /etc/profile.d/go_env.sh <<\EOF
-export GOPATH=/home/ubuntu/workspace/go
-export PATH=$GOPATH/bin:/usr/local/go/bin:$PATH
-EOF
-chmod +x /etc/profile.d/go_env.sh
-
-source /etc/profile.d/go_env.sh
-
-cat /vagrant/vimrc >> /etc/vim/vimrc
-
 export UCF_FORCE_CONFFNEW=YES
 export DEBIAN_FRONTEND=noninteractive
 
+VM_USER='ubuntu'
+DISK_DEVICE='/dev/disk/by-path/pci-0000:00:14.0-scsi-0:0:2:0'
+WORKSPACE="/home/${VM_USER}/workspace"
+DOCKER_DIR="${WORKSPACE}/docker_state"
+
+main() {
+  setup_workspace_disk
+  install_devtools
+  setup_golang
+  get_k8s_go_deps
+
+  cat /vagrant/vimrc >> /etc/vim/vimrc
+}
+
+setup_workspace_disk() {
+  echo "About to set up the workspace disk"
+  sudo -u "$VM_USER" mkdir -p "$WORKSPACE"
+
+  if ! mount | grep -q "$WORKSPACE"
+  then
+    if ! mount "$DISK_DEVICE" " $WORKSPACE " >/dev/null 2>&1 ; then
+      echo "couldn't mount disk, so making a fresh FS"
+      mkfs.ext4 "$DISK_DEVICE"
+      mount "$DISK_DEVICE" "$WORKSPACE"
+      chown "${VM_USER}.${VM_USER}" "$WORKSPACE"
+    fi
+  fi
+
+  mkdir -p "${DOCKER_DIR}"
+  ln -s "${DOCKER_DIR}" /var/lib/docker || true
+}
+
+install_devtools() {
+  apt-get -y update
+  apt-get -y clean
+
+  apt-get install -y git vim-nox jq cgroup-lite build-essential ntp htop docker.io
+}
+
+setup_golang() {
+  wget -qO- https://redirector.gvt1.com/edgedl/go/go1.9.2.linux-amd64.tar.gz \
+    | tar -C /usr/local -xzf -
+
+  # Set up $GOPATH and add go executables to $PATH
+  cat > /etc/profile.d/go_env.sh <<EOF
+  export GOPATH=/home/${VM_USER}/workspace/go
+  export PATH=\$GOPATH/bin:/usr/local/go/bin:\$PATH
+EOF
+  chmod +x /etc/profile.d/go_env.sh
+}
+
+get_k8s_go_deps() {
+  echo "Getting k8s golang dependences"
+  source /etc/profile.d/go_env.sh
+  CGO_ENABLED=0 go install -a -installsuffix cgo std
+
+  sudo -u "$VM_USER" -E bash <<'EOF'
+    source /etc/profile.d/go_env.sh
+    go get -u github.com/jteeuwen/go-bindata/go-bindata
+EOF
+
+}
+
+main
